Create separate packages for network and mocked tests
