+++ msal/src/test/java/com/microsoft/identity/client/robolectric/tests/mocked/AcquireTokenMockBaseTest.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.client.robolectric.tests;
+package com.microsoft.identity.client.robolectric.tests.mocked;
 
 import android.app.Activity;
 import android.content.Context;
@@ -30,30 +30,17 @@ import androidx.test.core.app.ApplicationProvider;
 import com.microsoft.identity.client.IPublicClientApplication;
 import com.microsoft.identity.client.PublicClientApplication;
 import com.microsoft.identity.client.exception.MsalException;
-import com.microsoft.identity.internal.testutils.labutils.TestConfigurationHelper;
-import com.microsoft.identity.internal.testutils.labutils.TestConfigurationQuery;
 
 import org.mockito.Mockito;
 
 import java.io.File;
 
-public abstract class AcquireTokenNetworkBaseTest {
+public abstract class AcquireTokenMockBaseTest {
 
     private static final String AAD_CONFIG_FILE_PATH = "src/test/res/raw/aad_test_config.json";
-    private static final String B2C_CONFIG_FILE_PATH = "src/test/res/raw/b2c_test_config.json";
 
-    private static final String AAD_AUTHORITY_TYPE_STRING = "AAD";
-    private static final String B2C_AUTHORITY_TYPE_STRING = "B2C";
-
-    /**
-     * @param publicClientApplication instance of Public Client Application
-     * @param activity                activity required for acquire token parameters
-     * @param username                username needed to attach to token request for ROPC
-     * @throws InterruptedException
-     */
     abstract void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
-                                       final Activity activity,
-                                       final String username) throws InterruptedException;
+                                       final Activity activity) throws InterruptedException;
 
 
     private Activity getActivity(final Context context) {
@@ -64,21 +51,17 @@ public abstract class AcquireTokenNetworkBaseTest {
     }
 
 
-    public void performTest(String authorityType) {
+    public void performTest() {
         final Context context = ApplicationProvider.getApplicationContext();
         final Activity testActivity = getActivity(context);
 
-        final String configFilePath = getConfigFilePath(authorityType);
-        final File configFile = new File(configFilePath);
-
-        final TestConfigurationQuery query = getTestConfigurationQuery(authorityType);
-        final String username = TestConfigurationHelper.getUpnForTest(query);
+        final File configFile = new File(AAD_CONFIG_FILE_PATH);
 
         PublicClientApplication.create(context, configFile, new PublicClientApplication.ApplicationCreatedListener() {
             @Override
             public void onCreated(IPublicClientApplication application) {
                 try {
-                    makeAcquireTokenCall(application, testActivity, username);
+                    makeAcquireTokenCall(application, testActivity);
                 } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
@@ -91,31 +74,4 @@ public abstract class AcquireTokenNetworkBaseTest {
         });
     }
 
-    private String getConfigFilePath(String authorityType) {
-        return authorityType == B2C_AUTHORITY_TYPE_STRING
-                ? B2C_CONFIG_FILE_PATH
-                : AAD_CONFIG_FILE_PATH;
-    }
-
-    /**
-     * @param authorityType can be either "AAD" or "B2C"
-     * @return test configuration query to be used for pulling test accounts from Lab Api
-     */
-    private TestConfigurationQuery getTestConfigurationQuery(String authorityType) {
-        return (authorityType == B2C_AUTHORITY_TYPE_STRING) ? getQueryForB2C() : getQueryForAAD();
-    }
-
-    private TestConfigurationQuery getQueryForAAD() {
-        final TestConfigurationQuery query = new TestConfigurationQuery();
-        query.userType = "Member";
-        query.isFederated = false;
-        query.federationProvider = "ADFSv4";
-        return query;
-    }
-
-    private TestConfigurationQuery getQueryForB2C() {
-        final TestConfigurationQuery query = new TestConfigurationQuery();
-        query.b2cProvider = "Local";
-        return query;
-    }
 }
