+++ msal/src/test/java/com/microsoft/identity/client/robolectric/tests/mocked/AcquireTokenMockTest.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.client.robolectric.tests;
+package com.microsoft.identity.client.robolectric.tests.mocked;
 
 import android.app.Activity;
 
@@ -37,6 +37,8 @@ import com.microsoft.identity.client.robolectric.shadows.ShadowMsalUtils;
 import com.microsoft.identity.client.robolectric.shadows.ShadowStorageHelper;
 import com.microsoft.identity.client.robolectric.shadows.ShadowStrategyResultServerError;
 import com.microsoft.identity.client.robolectric.shadows.ShadowStrategyResultUnsuccessful;
+import com.microsoft.identity.client.robolectric.utils.AcquireTokenTestHelper;
+import com.microsoft.identity.client.robolectric.utils.RoboTestUtils;
 import com.microsoft.identity.common.exception.ClientException;
 import com.microsoft.identity.common.internal.cache.ICacheRecord;
 import com.microsoft.identity.common.internal.providers.oauth2.TokenResponse;
